# Replace double-ended-queue with denque