`denque` is slightly faster than `double-ended-queue`.

Before:

```
$ node perf.js 
10 promises: 3.303ms
100 promises: 1.804ms
1000 promises: 10.349ms
10000 promises: 209.485ms
100000 promises: 1103.597ms
1000000 promises: 21267.233ms
```

After:

```
$ node perf.js 
10 promises: 2.259ms
100 promises: 1.194ms
1000 promises: 8.490ms
10000 promises: 196.285ms
100000 promises: 1080.001ms
1000000 promises: 19734.413ms
```