@lpinca 2017-06-06
The solution proposed in https://github.com/ForbesLindesay/throat/pull/15#issuecomment-306436704 yields similar results:

```
$ node perf.js 
10 promises: 2.230ms
100 promises: 1.181ms
1000 promises: 7.933ms
10000 promises: 137.024ms
100000 promises: 1165.815ms
1000000 promises: 19232.243ms
```

so maybe it's probably better to use that (no dependencies, less code).