@ForbesLindesay 2017-06-13
Thanks for these pull requests, I went with the other one.